avance

+++ Controllers/DepartamentosController.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Http;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using U3Api.Models.DTOs;
 using U3Api.Models.Entities;
@@ -9,6 +10,7 @@ namespace Proyecto_U3.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
+    [Authorize]
     public class DepartamentosController : ControllerBase
     {
         public Repository<Departamentos> Repository { get; }
@@ -18,78 +20,126 @@ namespace Proyecto_U3.Controllers
             Repository = repository;
         }
 
+
         [HttpGet]
         public IActionResult GetAllDepartamentos()
         {
-            var deptos = Repository.GetAll().
-                            OrderBy(x => x.Id).
-                            Select(x => new DepartamentoDTO
-                            {
-                                Id = x.Id,
-                                NombreDepartamento = x.Nombre,
-                                Username = x.Username,
-                                Password = x.Password,
-                                IdSuperior = x.IdSuperior
-                            });
-            return Ok(deptos);
+
+            var deptIdClaim = User.Claims.FirstOrDefault(x => x.Type == "id");
+
+            if (deptIdClaim == null || !int.TryParse(deptIdClaim.Value, out int deptid))
+            {
+                return Unauthorized();
+            }
+
+            var departamento = Repository.Get(deptid);
+
+            if(departamento.IdSuperior == null)
+            {
+                var deptos = Repository.GetAll().
+                          OrderBy(x => x.Id).
+                          Select(x => new DepartamentoDTO
+                          {
+                              Id = x.Id,
+                              NombreDepartamento = x.Nombre,
+                              Username = x.Username,
+                              Password = x.Password,
+                              IdSuperior = x.IdSuperior
+                          });
+                return Ok(deptos);
+            }
+
+            return Unauthorized("No eres administrador");
+          
         }
 
         [HttpPost]
         public IActionResult Post(DepartamentoDTO dto)
         {
-            DepartamentoValidator validator = new();
 
-            var resultados = validator.Validate(dto);
+            var deptIdClaim = User.Claims.FirstOrDefault(x => x.Type == "id");
 
-            if (resultados.IsValid)
+            if (deptIdClaim == null || !int.TryParse(deptIdClaim.Value, out int deptid))
             {
-                Departamentos entity = new()
-                {
-                    Id = 0,
-                    Nombre = dto.NombreDepartamento,
-                    Username = dto.Username,
-                    Password = dto.Password,
-                    IdSuperior = dto.IdSuperior
-                }; 
-
-                Repository.Insert(entity);
-                return Ok();
+                return Unauthorized();
             }
 
-            return BadRequest(resultados.Errors.Select(x => x.ErrorMessage));
+            var departamento = Repository.Get(deptid);
+
+
+            if (departamento.IdSuperior == null)
+            {
+                DepartamentoValidator validator = new();
+
+                var resultados = validator.Validate(dto);
+
+                if (resultados.IsValid)
+                {
+                    Departamentos entity = new()
+                    {
+                        Id = 0,
+                        Nombre = dto.NombreDepartamento,
+                        Username = dto.Username,
+                        Password = dto.Password,
+                        IdSuperior = dto.IdSuperior
+                    };
+
+                    Repository.Insert(entity);
+                    return Ok();
+                }
+
+                return BadRequest(resultados.Errors.Select(x => x.ErrorMessage));
+            }
+
+            return Unauthorized("No eres administrador");
         }
 
 
         [HttpPut("{id}")]
         public IActionResult Put(DepartamentoDTO dto)
         {
-            DepartamentoValidator validator = new();
 
-            var resultados = validator.Validate(dto);
+            var deptIdClaim = User.Claims.FirstOrDefault(x => x.Type == "id");
 
-            if (resultados.IsValid)
+            if (deptIdClaim == null || !int.TryParse(deptIdClaim.Value, out int deptid))
             {
-                var entidadDept = Repository.Get(dto.Id ?? 0);
+                return Unauthorized();
+            }
 
-                if (entidadDept == null)
+            var departamento = Repository.Get(deptid);
+
+            if (departamento.IdSuperior == null)
+            {
+                DepartamentoValidator validator = new();
+
+                var resultados = validator.Validate(dto);
+
+                if (resultados.IsValid)
                 {
-                    return NotFound();
-                }
-                else
-                {
-                    entidadDept.Nombre = dto.NombreDepartamento;
-                    entidadDept.Username = dto.Username;
-                    entidadDept.Password = dto.Password;
-                    entidadDept.IdSuperior = dto.IdSuperior;
+                    var entidadDept = Repository.Get(dto.Id ?? 0);
 
-                    Repository.Update(entidadDept);
+                    if (entidadDept == null)
+                    {
+                        return NotFound();
+                    }
+                    else
+                    {
+                        entidadDept.Nombre = dto.NombreDepartamento;
+                        entidadDept.Username = dto.Username;
+                        entidadDept.Password = dto.Password;
+                        entidadDept.IdSuperior = dto.IdSuperior;
 
-                    return Ok();
+                        Repository.Update(entidadDept);
+
+                        return Ok();
+                    }
                 }
+
+                return BadRequest(resultados.Errors.Select(x => x.ErrorMessage));
 
             }
 
-            return BadRequest(resultados.Errors.Select(x => x.ErrorMessage));
+            return Unauthorized("No eres administrador");
         }
 
 
@@ -103,9 +153,25 @@ namespace Proyecto_U3.Controllers
                 return NotFound();
             }
 
-            Repository.Delete(entidadDept);
 
-            return Ok();
+            var deptIdClaim = User.Claims.FirstOrDefault(x => x.Type == "id");
+
+            if (deptIdClaim == null || !int.TryParse(deptIdClaim.Value, out int deptid))
+            {
+                return Unauthorized();
+            }
+
+            var departamento = Repository.Get(deptid);
+
+            if (departamento.IdSuperior == null)
+            {
+                Repository.Delete(entidadDept);
+
+                return Ok();
+            }
+
+            return Unauthorized("No eres administrador");
+
         }
     }
 }
